rewrite how graph construction works to not require clone all the time ;)

+++ src/cell.rs
@@ -3,7 +3,7 @@ use crate::vector_3d::Vector3D;
 #[derive(Hash, PartialEq, Clone, Debug)]
 pub struct Cell {
     pub center: Vector3D,
-    pub label: i32,
+    pub label: usize,
 }
 
 impl Eq for Cell {
